requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
openai
python-multipart
playwright